chore: move careers, pricing, privacy, status and terms into main folder for better organization and structure
